Improves socket permission. (#84)
* Fixes socket permission for handle sig.
